refactor: extract shared validation, eliminate `as any`, name magic constants
- Extract isTarget/isTargetsRule to shared targets-rule.ts, removing
  identical code duplicated across round-robin/random/failover strategies
- Replace `as any` with typed TargetInput interface in groups.ts
- Extract FAILOVER_FAIL_THRESHOLD and MIN_FAILOVER_TARGETS constants
