+++ src/proxy/strategy/round-robin.ts
@@ -1,25 +1,5 @@
 import type { MappingStrategy, ResolveContext, Target } from "./types.js";
-
-interface TargetsRule {
-  targets: Target[];
-}
-
-function isTarget(value: unknown): value is Target {
-  return (
-    typeof value === "object" &&
-    value !== null &&
-    "backend_model" in value &&
-    typeof (value as Target).backend_model === "string" &&
-    "provider_id" in value &&
-    typeof (value as Target).provider_id === "string"
-  );
-}
-
-function isTargetsRule(value: unknown): value is TargetsRule {
-  if (typeof value !== "object" || value === null) return false;
-  const r = value as TargetsRule;
-  return Array.isArray(r.targets) && r.targets.every(isTarget);
-}
+import { isTargetsRule } from "./targets-rule.js";
 
 export class RoundRobinStrategy implements MappingStrategy {
   private indexMap = new Map<string, number>();
