+++ src/proxy/strategy/targets-rule.ts
@@ -1,0 +1,22 @@
+import type { Target } from "./types.js";
+
+interface TargetsRule {
+  targets: Target[];
+}
+
+export function isTarget(value: unknown): value is Target {
+  return (
+    typeof value === "object" &&
+    value !== null &&
+    "backend_model" in value &&
+    typeof (value as Target).backend_model === "string" &&
+    "provider_id" in value &&
+    typeof (value as Target).provider_id === "string"
+  );
+}
+
+export function isTargetsRule(value: unknown): value is TargetsRule {
+  if (typeof value !== "object" || value === null) return false;
+  const r = value as TargetsRule;
+  return Array.isArray(r.targets) && r.targets.every(isTarget);
+}
